Change f32 to f64 and start on having separate fields

+++ engine_common/src/engine_bus.rs
@@ -6,20 +6,20 @@ use system_interface::common::keyboard_interface::{MyKeyboardEvent};
 pub enum ForceField{
     Constant{ // e.g. gravity on a 2D sidescroller
         force : Vector2,
-        charge : Box<[f32]>        
+        charge : Box<[f64]>        
     },
     InverseSquare{ // e.g. gravity in a space sim
         source : usize,
-        constant : f32,
-        charge : Box<[f32]>
+        constant : f64,
+        charge : Box<[f64]>
     },
     InverseSquareNBody{// e.g. N body gravity sim
-        constant : f32,
-        charge : Box<[f32]>
+        constant : f64,
+        charge : Box<[f64]>
     },
     LJPotential{ // Leonard-Jones potential, used for simulating phases of matter
-        a : f32,
-        b : f32
+        a : f64,
+        b : f64
     }
 }
 
@@ -30,15 +30,15 @@ pub enum LogicToPhysicsChannel{
     },
     AddEntity{
         group_index : usize,
-        mass : f32,
-        moi : f32,
-        posx : f32,
-        posy : f32,
-        velx : f32,
-        vely : f32,
-        orien : f32,
-        angvel : f32,
-        shape : Vec<f32>
+        mass : f64,
+        moi : f64,
+        posx : f64,
+        posy : f64,
+        velx : f64,
+        vely : f64,
+        orien : f64,
+        angvel : f64,
+        shape : Vec<f64>
     },
     RemoveEntity{
         group_index : usize,
@@ -47,18 +47,18 @@ pub enum LogicToPhysicsChannel{
     ApplyForce{
         group_index : usize,
         index : usize,
-        forcex : f32,
-        forcey : f32
+        forcex : f64,
+        forcey : f64
     },
     ApplyCenterlineForce{
         group_index : usize,
         index : usize,
-        force : f32
+        force : f64
     },
     ApplyTorque{
         group_index : usize,
         index : usize,
-        torque : f32
+        torque : f64
     },
     ApplyField{
         group_index : usize,
